requirements.txt
# DICOM Viewer - Python Dependencies
# Copyright (c) 2026 Divergent Health Technologies
#
# This file lists Python packages required for the Flask backend.
# Install with: pip install -r requirements.txt
# Recommended: Use a virtual environment (python -m venv venv)

flask==3.1.2        # Web framework - serves static files and provides test mode API
pydicom==3.0.1      # DICOM file parsing - used for server-side test data scanning
PyJWT==2.9.0        # JWT token generation and validation for cloud auth
